requirements.txt
fastapi
uvicorn
transformers
pydantic
pandas
scikit-learn
torch
passlib
bcrypt==3.1.7
PyJWT>=2.0.0
python-multipart
streamlit
pyjwt


